Started adding upgrade downloading. Still completely untested.

+++ src/uk/digitalsquid/netspoofer/InstallService.java
@@ -37,6 +37,8 @@ import java.util.concurrent.CancellationException;
 import java.util.zip.GZIPInputStream;
 
 import uk.digitalsquid.netspoofer.config.Config;
+import uk.digitalsquid.netspoofer.config.ConfigChecker;
+import uk.digitalsquid.netspoofer.misc.XDelta;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
@@ -58,6 +60,8 @@ public class InstallService extends Service implements Config {
 	 */
 	public static final String INTENT_START_FILE = "uk.digitalsquid.netspoofer.config.InstallStatus.isFile";
 	public static final String INTENT_START_URL_UNZIPPED = "uk.digitalsquid.netspoofer.config.InstallStatus.URLUnzipped";
+	public static final String INTENT_START_URL_UPGRADE = "uk.digitalsquid.netspoofer.config.InstallStatus.isUpgrade";
+	public static final String INTENT_START_KEEP_INSTALLATION_FILE = "uk.digitalsquid.netspoofer.config.InstallStatus.keepInstallationFile";
 	public static final int STATUS_STARTED = 0;
 	public static final int STATUS_DOWNLOADING = 1;
 	public static final int STATUS_FINISHED = 2;
@@ -115,11 +119,13 @@ public class InstallService extends Service implements Config {
 		// String downloadUrl = prefs.getString("debImgUrl", DEB_IMG_URL);
 		String downloadUrl = intent.getStringExtra(INTENT_START_URL);
 		boolean downloadUnzipped = intent.getBooleanExtra(INTENT_START_URL_UNZIPPED, false);
+		boolean isUpgrade = intent.getBooleanExtra(INTENT_START_URL_UPGRADE, false);
 		boolean useLocalFile = intent.getBooleanExtra(INTENT_START_FILE, false);
+		boolean keepInstallationFile = intent.getBooleanExtra(INTENT_START_FILE, false);
 		if(downloadUrl == null) throw new IllegalArgumentException("Start URL was null");
 		Log.v(TAG, "Downloading file " + downloadUrl);
 		// if(downloadUrl.equals("")) downloadUrl = DEB_IMG_URL;
-		downloadTask.execute(new DlStartData(downloadUrl, downloadUnzipped, useLocalFile));
+		downloadTask.execute(new DlStartData(downloadUrl, downloadUnzipped, isUpgrade, useLocalFile, keepInstallationFile));
 	}
 	
 	/**
@@ -131,12 +137,23 @@ public class InstallService extends Service implements Config {
 		private static final long serialVersionUID = 6287320665354658386L;
 		public final String url;
 		public final boolean unzipped;
+		/**
+		 * When <code>true</code>, indicates an incremental upgrade (ie. a patch).
+		 */
+		public final boolean upgrade;
 		public final boolean useLocalFile;
 		
-		public DlStartData(String url, boolean unzipped, boolean useLocal) {
+		/**
+		 * When <code>true</code>, the gz file is kept for use in an upgrade.
+		 */
+		public final boolean keepInstallationFile;
+		
+		public DlStartData(String url, boolean unzipped, boolean isUpgrade, boolean useLocal, boolean keepInstallationFile) {
 			this.url = url;
 			this.unzipped = unzipped;
+			this.upgrade = isUpgrade;
 			useLocalFile = useLocal;
+			this.keepInstallationFile = keepInstallationFile;
 		}
 	}
 	
@@ -185,11 +202,6 @@ public class InstallService extends Service implements Config {
 			this.status = status;
 		}
 		
-		@Deprecated
-		public boolean isExtracting() {
-			return getStatus() == STATUS_DOWNLOADING;
-		}
-		
 		public int getBytesDone() {
 			return bytesDone;
 		}
@@ -228,10 +240,13 @@ public class InstallService extends Service implements Config {
 		private URLConnection connection;
 		
 		private File sd;
-		private File debian;
+		/**
+		 * This is the destination to download to. In the case of an upgrade, it is the upgrade file.
+		 */
+		private File dlDestination;
 		
 		private URL downloadURL;
-		private FileOutputStream debWriter;
+		private FileOutputStream dlWriter;
 		
 		public void publishDLProgress(DLProgress progress) {
 			publishProgress(progress);
@@ -239,32 +254,45 @@ public class InstallService extends Service implements Config {
 		
 		@Override
 		protected Integer doInBackground(DlStartData... params) {
+			if(params.length != 1) throw new IllegalArgumentException("Please specify 1 parameter");
 			boolean downloadUnzipped = params[0].unzipped;
 			boolean useLocalFile = params[0].useLocalFile;
+			final boolean upgrade = params[0].upgrade;
+			final boolean keepInstallationFile = params[0].keepInstallationFile;
+			if(upgrade) { // Never use local file when upgrading, and don't extract
+				useLocalFile = false;
+				downloadUnzipped = true;
+			}
 			sd = getExternalFilesDir(null);
-			debian = downloadUnzipped ?
-					new File(sd.getAbsolutePath() + "/" + DEB_IMG) : // Save directly to new location
-					new File(sd.getAbsolutePath() + "/" + DEB_IMG_GZ);
+			
+			if(upgrade) {
+				dlDestination = new File(getFilesDir() + "/" + "upgrade.vcdiff");
+			}
+			if(downloadUnzipped) {
+				dlDestination = new File(sd.getAbsolutePath() + "/" + DEB_IMG); // Save directly to new location
+			} else {
+				dlDestination = new File(sd.getAbsolutePath() + "/" + DEB_IMG_GZ);
+			}
 			try {
-				debian.createNewFile();
+				dlDestination.createNewFile();
 			} catch (IOException e) {
 				e.printStackTrace();
 				return STATUS_DL_FAIL_SDERROR;
 			}
 			
 			// Delete version file
+			final int oldVersionNumber = ConfigChecker.getVersionNumber(getBaseContext());
 			File oldversion = new File(sd.getAbsolutePath() + "/" + DEB_VERSION_FILE);
 			oldversion.delete();
 			
 			boolean done = true;
 			if(!useLocalFile) {
 				try {
-					debWriter = new FileOutputStream(debian);
+					dlWriter = new FileOutputStream(dlDestination);
 				} catch (FileNotFoundException e) {
 					e.printStackTrace();
 				}
 				
-				if(params.length != 1) throw new IllegalArgumentException("Please specify 1 parameter");
 				try {
 					downloadURL = new URL(params[0].url);
 				} catch (MalformedURLException e) {
@@ -298,22 +326,44 @@ public class InstallService extends Service implements Config {
 				}
 				
 				try {
-					debWriter.close();
+					dlWriter.close();
 					if(response != null) response.close(); // Could be nothing that was downloaded
 				} catch (IOException e) {
 				}
 			} else { // Use local
 				File localFile = new File(params[0].url);
 				if(!localFile.exists()) return STATUS_DL_FAIL_IOERROR;
-				if(!localFile.renameTo(debian)) return STATUS_DL_FAIL_SDERROR;
+				if(!localFile.renameTo(dlDestination)) return STATUS_DL_FAIL_SDERROR;
 			}
 			
 			if(!downloadUnzipped) { // Don't bother extracting
 				try {
-					unzip(debian);
+					unzip(dlDestination, !keepInstallationFile);
 				} catch (IOException e1) {
 					e1.printStackTrace();
 					done = false;
+				}
+			}
+			
+			if(upgrade) { // Perform XDelta based upgrade
+				try {
+					if(!XDelta.patchFile(this, getBaseContext(), dlDestination.getAbsolutePath(), DEB_IMG, DEB_IMG_GZ))
+						throw new IOException("File patching failed");
+				} catch (IOException e) {
+					Log.w(TAG, "File patch failed", e);
+					done = false;
+					File version = new File(sd.getAbsolutePath() + "/" + DEB_VERSION_FILE);
+					try {
+						version.createNewFile(); // Make sure exists
+						FileOutputStream versionWriter = new FileOutputStream(version);
+						versionWriter.write(("" + oldVersionNumber).getBytes()); // Write previous version
+						versionWriter.close();
+					} catch (FileNotFoundException e1) {
+						e1.printStackTrace();
+					} catch (IOException e1) {
+						e1.printStackTrace();
+						Log.e(TAG, "Couldn't write version file");
+					}
 				}
 			}
 			
@@ -359,7 +409,7 @@ public class InstallService extends Service implements Config {
 			try {
 				while((bytesRead = response.read(dlData)) != -1) {
 					bytesDone += bytesRead;
-					debWriter.write(dlData, 0, bytesRead);
+					dlWriter.write(dlData, 0, bytesRead);
 					if(i++ > 6) {
 						i = 0;
 						progress.setBytesDone(bytesDone + bytesSoFar);
@@ -368,7 +418,7 @@ public class InstallService extends Service implements Config {
 					
 					if(isCancelled()) {
 						response.close();
-						debWriter.close();
+						dlWriter.close();
 						// Remove old files
 						File delversion = new File(sd.getAbsolutePath() + "/" + DEB_VERSION_FILE);
 						delversion.delete();
@@ -388,7 +438,7 @@ public class InstallService extends Service implements Config {
 			return bytesDone;
 		}
 		
-		private String unzip(File inFile) throws IOException
+		private String unzip(File inFile, boolean deleteOld) throws IOException
 		{
 		    InputStream gzipInputStream = new BufferedInputStream(new GZIPInputStream(new FileInputStream(inFile)));
 		 
@@ -413,7 +463,7 @@ public class InstallService extends Service implements Config {
 		    gzipInputStream.close();
 		    out.close();
 		 
-		    inFile.delete();
+		    if(deleteOld) inFile.delete();
 		 
 		    return outFilePath;
 		}
